move SourceContext tests inside TextInput.spec
